Fix not setting required header fields after message processing
+ adjusting connection parameters (chunking, keepalive) according
to plugins-generated header

response body data caching for late processing is WIP

+++ src/sk/fiit/peweproxy/utils/InMemBytesStore.java
@@ -11,7 +11,7 @@ public class InMemBytesStore {
 	private final ByteArrayOutputStream stream;
 	private final Queue<Integer> dataIncrements;
 	
-	public InMemBytesStore(int initSize) {
+	public InMemBytesStore(long initSize) {
 		if (initSize < 1) {
 			initSize = DEF_ARRAYSIZE;
 		}
